rename chunk.py to disk_buffered_message.py
